Sort results of output from the "find" command
Filesystems do not all guarantee that 2 successive invocations of find
will return results in the same order. It therefore makes sense to sort
the results.  This patch attempts to sort only if the comparison of 2
file lists fails without it. Therefore the penalty of sorting is only paid
on filesystems which do not return results in the same order every time.
In this way we avoid regenerating the ninja file at times when there is
no relevant reason to do so.

Change-Id: I357eb2f139f7a7f04f039090b603382f08d1a576